requirements.txt
discord.py==2.3.2
yt-dlp==2023.12.30
PyNaCl==1.5.0
python-dotenv==1.0.0
fastapi==0.109.2
uvicorn==0.27.1 